requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
itsdangerous==2.1.2
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
azure-communication-sms==1.0.1
WTForms==3.1.1
Flask-WTF==1.2.1
email-validator==2.1.0